pyproject.toml

[tool.poetry]
name = "source"
version = "0.1.0"
description = "Cookiecutter machine learning project repository"


authors = ["Julian Cooper"]

readme = "README.md"
repository = "https://github.com/bcgx-pi-digitaltravel/ba-ac-availability-prediction"


[tool.poetry.dependencies]
python = "^3.8.1"

more-itertools = "^8.13.0"
matplotlib = "^3.5.2"
ipython = "^8.4.0"
numpy = "^1.21"
pandas = "^1.3"
plotly = "^5.11.0"
pyyaml = "^6.0"
xgboost = ">=2.0.1"


[tool.poetry.group.dev.dependencies]
black = ">=22.8"
flake8 = ">=5.0"
furo = ">=2023.9.10"
isort = ">=5.10"
pylint = ">=2.15"
pytest = ">=7.1"
pytest-cov = ">=3.0"
sphinx = ">=5.1"
sphinxcontrib-napoleon = ">=0.7"
myst-parser = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pylint.main]
# Specify a score threshold under which the program will exit with error.
fail-under = 5
# Files or directories to be skipped. They should be base names, not paths.
ignore = [".git", "__pycache__", ".pytest_cache", "docs", "venv"]
# Discover python modules and packages in the file system subtree.
recursive = true
# Use multiple processes to speed up Pylint. Specifying 0 will auto-detect the
# number of processors available to use, and will cap the count on Windows to
# avoid hangs.
jobs = 0

[tool.pylint."messages control"]
disable = ["R", "C", "W"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)7s] [%(filename)s:%(lineno)-3s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
